Organize classes, and move to the admin folder.
