+++ src/Tickets/Emails/Provider.php
@@ -36,7 +36,7 @@ class Provider extends tad_DI52_ServiceProvider {
 		// Register singletons.
 		$this->container->singleton( static::class, $this );
 		$this->container->singleton( 'tickets.emails.provider', $this );
-		$this->container->singleton( Emails_Tab::class );
+		$this->container->singleton( Admin\Emails_Tab::class );
 		$this->container->singleton( 'tickets.emails.emails-tab', $this );
 
 		$this->container->singleton( Admin\Preview_Modal::class );
